refactor(install): split config modules
